% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tensortoeplitz}
\alias{tensortoeplitz}
\title{An RcppArmadillo function to compute the
inverse of a tensor decomposed Toeplitz matrix. (b*I_{nd} + J_n tensor A)
For toeplitz A}
\usage{
tensortoeplitz(b, n, c)
}
\arguments{
\item{b}{The magnitude of the ridge added to the tensor decomposition}

\item{n}{an Integer of the size of the tensor decomposition}

\item{c}{A numeric vector, which is the first row of the Toeplitz matrix A}
}
\value{
The inverse of the tensor decomposed Toeplitz matrix
}
\description{
An RcppArmadillo function to compute the
inverse of a tensor decomposed Toeplitz matrix. (b*I_{nd} + J_n tensor A)
For toeplitz A
}
