\name{toeplitz-package}
\alias{toeplitz-package}
\alias{toeplitz}
\docType{package}
\title{
  An RcppArmadillo package for accelerated linear algebra with Toeplitz matrices
}
\description{
 An RcppArmadillo package for high-performance linear algebra with Toeplitz matrices.
  Trench and Durbin algorithm are written in c++ and linked to R.
}
\details{
 There are function to compute the inverse and the determinant of a Toeplitz matrix.
}
\author{
Oliver M. Crook

Maintainer: Oliver M. Crook <omc25@cam.ac.uk>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
