% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trenchDetcpp}
\alias{trenchDetcpp}
\title{An RcppArmadillo function to compute the
determinant of a Toeplitz matrix}
\usage{
trenchDetcpp(c)
}
\arguments{
\item{c}{A numeric vector, which is the first row of the Toeplitz matrix}
}
\value{
A list containing the log determinant and two vectors z and v
to be passed to trenchInv to calculate the inverse.
}
\description{
An RcppArmadillo function to compute the
determinant of a Toeplitz matrix
}
