% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trenchInvcpp}
\alias{trenchInvcpp}
\title{An RcppArmadillo function to compute the
inverse of a Toeplitz matrix}
\usage{
trenchInvcpp(v)
}
\arguments{
\item{v}{A numeric vector, which is calculated from trenchDetcpp}
}
\value{
The inverse of the toeplitz matrix
}
\description{
An RcppArmadillo function to compute the
inverse of a Toeplitz matrix
}
